Day 20 - Pulses - Part 2
Needed to look internally to the button presses to find
when each of the inputs to the Conjunctions were high.
Apparently they work nicely and always loop evenly.
Figure out when the Conjunction in front of RX ("GQ") is
high, because then it emits low to RX and we win.