2022 Day 15 - Part 2
Using a modified version of Todd's IntRange reduction,
we can turn each row into one or two IntRanges.
Find the one row with two ranges, and the X coordinate
is the value in between the ranges.
Takes a while to run because it is semi-brute force,
but finishes in under 20 seconds.