2024 Day 4
Search for the key letter, then look at the neighbors.

Part one we find X and then look at all possible paths
for the next three letters.  If they are in the grid
and match the correct letters, we count it.

Part two we find A and look at the four valid patterns
of M and S to see if count the A.