2024 Day 5
Make data classes for the Manual and a Page for code organization.
Each Page is responsible for looking at the rules to see if it
is valid.  Compare each page number to all following page numbers.
If we find a rule equal to the second page | first page,
the page is invalid.

Part 1 we find the valid pages and add the middle numbers.
Part 2 we find the invalid pages then fix them recursively.
To fix a page, find the first violated rule and swap the
offending numbers. If the resulting page is valid, return it.
If it is invalid, try to fix the new page.