2024 Day 6 - Part 2
Look at each space in the path (ignoring starting space).
Walk the path and pretend that space is an obstacle.
Track the path and the direction we're moving.
If at any point we repeat a location and direction, we're in a loop,
so we include that in our count.
If we get to the edge despite our new obstacle, we do not count it.