2024 Day 7
Turn each line into a data class containing the target and the
list of numbers. Go through the list of numbers trying
each operation recursively with the remainder of the list.
Part one does n^2 operations, part two does n^3.
When we have the list of operations, we can check if the
target is present.  Filter to the lines that are valid,
add their targets, and we are done.