2024 Day 8
Iterate through the input and make a map of character to list
of location. For each location list, make all possible pairs of
locations, and order does matter. For each pair of locations,
find the destination if we walk from the first to the second.
For part two, we return a list of destinations if we keep walking
as long as we stay in bounds. Combine the list of destinations
together as our possible antinodes.
In part two, we also have to add back in antenna locations
for any antenna that has more than one location, because that
is considered a valid antinode.
Check that the antinodes are in bounds for safety, use a set
to enforce uniqueness, and count the result.