2024 Day 9 - Part 2
Instead of a list of Block, use a list of list of Block.
This allows us to look for empty segments that can fit
our whole segment of blocks. Eventually we flatten the list
and reuse our checksum function. Update the checksum function
to handle Empty Blocks as 0 value instead of throwing.