2024 Day 10 - Part 1
Decided against BFS after some attempts.
Create two maps for efficiency.
Map one is Location to Set of Peak Locations reachable by walking.
The peaks list is initially empty except for peaks themselves.
Map two is Height to Set of Locations at that height to avoid
extra looping through the input.
Descending in height, look at each location of a given height.
For each of those locations, look for neighbors that are one step
above that location. Add their set of valid peaks to the current
location's set of peaks. Eventually we get down to the trailheads,
and they have a set of peaks they can each reach.
Sum the size of these sets and solve.