2024 Day 11 - Part 2
Inspired by Todd Ginsburg's solution to 2021 Day 6, we do not keep
track of individual stones. For each iteration, we store a count
of how many of a given stone number is present in a map.
We produce a new map each day by shifting stones to their new
value(s).  Use Map.merge() to set the value to either the count
or add the count to an existing value.