Day 12 - Part 1
First we sort every location by letter.
Then we divide each group of letters into plots of locations.
We calculate area by counting the locations in a plot.
We calculate perimeter by taking the number of sides per
location (4) and subtracting the number of neighbors
that are part of the same plot.

Ideally we would find a way to not pass input everywhere for
the indices.