Day 15 - Part 1
Create a warehouse class that stores the location of the robot,
and lists of the box and wall locations.
At each step, we see if the space we want to move into is empty.
If so, we can move the robot.
If it is a wall, we're done.
If it is a box, keep looking forward to see if we find a wall
or an empty space first. Use that to decide if we push the first
box into the empty spot, or if we remain still.