Day 16 - Part 1
DFS could not handle the larger map, so we needed inspiration
from Reddit. Walk in a straight line to try and find the exit.
Take note of every turn we could take.
If we did not find the exit, increment our turn count, then walk
in a straight line from each of the turns we noted from our last
set of straight lines. Keep track of locations we have turned from
and never try turning from there again.
Turns are so expensive that as soon as we find our first path
that reaches the exit, we know that's the best.