Day 17 - Part 2
Reddit guided answer again.  I knew we needed to go right to left,
but hadn't noticed the program is a loop.
The winning algorithm is find all possible inputs between 0 and 7
that produce the last instruction in the list.
Then, for each of those values, shift them over by modulo 8,
so bit-shift 3 to the left, and try adding 0 to 7 to them
to get a program that outputs the next instruction from the end.
Eventually we iterate through all instructions we need to
find, and we have a valid A register.