Day 18 - Part 2
Initially used brute force to run Dijkstra with every number
of rocks until we hit a failure. That worked but was slow.
Changed the findFirstBlocker method to binary search for
the first failing number of obstacles.
Remember to offset by one to convert from the failing number
of rocks to the index of the failing rock.