Day 19 - Part 1
Look at a string and see if it starts with a towel.
If so, look at the rest of the string and see if it starts
with a towel. If we get to an empty string, we have found a valid
combo because we have used towels to get there.
Cache all substring possibilities in a map so we don't duplicate
work (like the initial attempt).