Day 21 - Part 2
Changed the cache based on Todd's example, otherwise we OOM.
The realization is that after every pair of characters in the
top level code, all robots will be pointing back at A.
So we need to go to full depth for each pair of characters at the
top level, but not for the full sequence.

In hindsight, manually coding the directions is a buggy nightmare.
The biggest issue was an extra path between ^ and < that was
illegal, skewing my results.