Day 22 - Part 2
For each starting secret, we find the 2000 values we will produce.
Use windowed to go through each set of 5 values (4 changes)
and if it is the first time we've seen that sequence, add it
to a running map of sequences to bananas.
When we have finished every starting secret, we can pick the
largest number of bananas.