Day 24 - Part 2
The code is a bit all over the place and requires manual work.
The first step is to trace back the heritage of the gates to
the base x and y wires. For every gate, find the heritage
of the first wire and second wire, alphabetize the results and
wrap it in parentheses. Recurse this formula and you get a map
of wire name to gate string. Print out the Z gates and focus
on the first instance where the actual bit in the result doesn't
match the expected value. Manually figure out what gates to
swap to fix that gate, then move on to the next issue.
Swapping uses some convenient Set operations in a fold.
As we added more swaps, the problems got pushed to higher Z values
until addition works.