2015 Day 6
Convert each line into an instruction to be processed.

Part 1 we track a set of lit lights. For every instruction
we add to the set or remove depending on the on/off.
For toggle, we try to remove each light. If remove returns false,
the light was not on initially so we turn it on.

Part 2 we use a map of light location to brightness, then process
each light.