2015 Day 7
Use a sealed class to represent the types of wires.
The tricky part is that all types of gates could take raw values
or other wires as input.
Use a local function to get the value from the map or the raw
integer value per wire.
Part 1 we process the circuits and output the A wire.
Part 2 we take the result of part A, replace the original B wire
with a new wire that takes part 1's result as input,
then process the circuits a second time.