2015 Day 10
Loop through the characters of a string.  If the digit matches
the current digit we're using, increase our counter.
If it doesn't match, append the count and the digit to the result
string.  If we reach the end of the string, append the current
count and digit.
Call that method in a loop on the result for the required number
of times, then get the length of the result.