2015 Day 18
Track the lights that are on in the grid.
At each step, populate a new set of locations for the lights
that should be on.
For part 2, keep an extra set of the lights that always stay lit,
and use that as the basis for the new set of locations each time.