Year 2016 - Day 7
Split each line into the strings outside the brackets and the
strings inside the brackets.
Part 1 look for instances where the first list has a hit and
the second list has no hits.  Watch out for instances of
aaaa instead of abba.
Part 2 look for aba in the first list, but watch out for
overlapping matches.  Look for matches of the results in
the second list.