Year 2016 - Day 10
For each line, either add a value to a bot's list of values,
or create a function to execute when the bot has two values.
Loop until no bots have two values to process.
Track the combinations we've sorted and the outputs.
Return both in a data class so parts one and two can
look at what they need.