Year 2016 - Day 12
Store the 4 registers in a map for easy read/write access.
Take initial C value as input to handle part 1 vs 2.
Handle instructions one at a time.
Move the counter forward one after each instruction unless
we're doing a jump, in which case we override that offset.
Similar to 2025 day 23 but easier.