Year 2016 - Day 14
Store the calculated MD5s in a map using the index as the key,
that way we never recalculate a hash.
Use Map.getOrPut() to make that process simple.
Use a helper function to get a hash for a given index.
Optionally repeat the hashing process to support part 2.
Whenever we find a hash with 3 repeating characters,
look at the next 1000 hashes for the 5 character string.
This look ahead also serves to populate the map of hashes if
we're not there yet.