Year 2016 - Day 15
Convert the lines of the input into Predicates for when
the capsule can succeed on each line.
Reduce the list of Predicate into a single Predicate using
the and() function.
Generate a sequence of N to find the first N where the combined
Predicate returns true.
For Part 2 we simply append an extra line to our input and
run the function again.