Year 2016 - Day 18
Function to construct a maze by adding rows until we
hit the desired length.
Function to create the next line by looking at the current line
and enforcing the rules.  Simplify the 4 rules down to a single
rule that if Left != Right, we get a trap.
Function to count the safe spaces in the full maze.