Year 2016 - Day 19
Part 1 we realize that we can drop every elf with an even index.
If the overall list length was odd, the final elf effectively
becomes the first elf of the next iteration.
Use tailrec recursion to quickly shrink the list down to one.

Part 2 we use the base case of a single elf, then use induction
to solve for N+1 elves when we know the answer for N elves.