Year 2016 - Day 20
Convert each line into ranges.
For part 1, find the first number that isn't in any range.
We can brute force this because we eliminate numbers quickly.
For part 2, combine any overlapping or adjacent ranges.
Be sure to always take the highest upper limit, which means
some ranges are entirely subsumed by other ranges.
When we have our combined ranges, sum their sizes (remembering
to add 1 because they are inclusive), then subtract this sum
from the number of IPs (MAX IP + 1 because of 0).