Year 2016 - Day 22
Part 1 we can programmatically find all pairs of nodes,
filter to pairs that are distinct, then check if the available
space is greater than or equal to the used space.

Part 2 we print out the grid and find the only empty space.
Count the number of steps required to get to the top right,
noticing tha there is a wall of big numbers in the way.
After we shift the top right corner one time, we can see
that it takes 5 moves to shift the empty space in front of it
and move it left one time.  So the formula is
(# moves to initially move top right) + (5 * columns from top left)