Year 2016 - Day 23
Part 1 we update the code from day 12 and introduce the
new toggle option.  This runs quickly and is easy.
Part 2 we have to decompile the instructions to identify
the two main loops, which each have some nested loops in them.
Eventually we figure out a reasonable math problem to do on
a calculator, which I have recreated in the code.