Year 2016 - Day 24
Parse the input maze to get a set of valid locations.
Include the points of interest in the set of locations, but
also keep track of those POIs.
Run Dijkstra from each POI and store the results in a map of maps.
Starting at 0, do a BFS going to each POI that hasn't been
visited yet, adding the distances and taking the minimum path.
For part 2, we only have to change the base case for when
there are no remaining points to visit. Instead of adding 0
steps to the running total, we add the distance from the final
point back to home.