Year 2025 - Day 01
Start by processing the list of directions to produce
a running list of absolute positions.

Part 1, we see if the final position after a rotation
is at 0, and count those instances.

Part 2, we look at the range we cover for a given move
and count how many interim positions are 0.
To avoid double counting, we exclude our starting position
because it is already counted as a finishing position.