Year 2025 - Day 03
Initial implementation for two steps was more hardcoded.
Part 2 was easier with recursion.

Find the largest digit that has enough digits after it
to make a valid answer.
Then search the substring after the first instance of that
digit to find the largest digit again.
Repeat until the accumulated string is the desired size.