Year 2025 - Day 06
Part 1 we split the lines and trim out any blanks.
Iterate through the operators with an index, using that index
to get the numbers from the first 3 lines.
Accumulate the sum of the operated results.

Part 2 we loop backwards from the maximum index to get the
right to left behavior.  Pad each line out to the max length.
Store the numbers we create until we hit an operator,
use the operator, update a running sum, and clear our
list of numbers to be reused again.
Eventually return the grand sum.