Year 2025 - Day 07
Part 1 we track which indexes have a beam.
Use a set to avoid duplicated work.
Increment a counter each time we perform a split and
return that counter in the end.

Part 2 we track how many beams are in each index.
Use a map and add our counts as we go, even if we are not
doing a split, because another beam could split into us.
Sum the counts at the end.