Year 2025 - Day 09 - Part 2
Credit to Todd for the insight.
I knew that each line between corners is an edge.
The insight is to treat each edge as a 1 by N rectangle.
Next, create all rectangles between corners that we want to
test. Sort descending by area and filter until we find one
that does not leave the bounds.

To check if a rectangle is in bounds, validate that other than
its border, it never intersects an edge rectangle.
Todd's way of doing this is to take the inside of our
candidate rectangle and see if that intersects with any edge.
If it does, then the full size rectangle would have passed
beyond the bounds and is invalid.