Year 2025 - Day 10 - Part 1
Parse each line into a 3 part machine.
Represent the indicator lights as booleans.
Represent the buttons as lists of Int.

Starting at groups of 1 button and increasing in size,
check if any group of that size is valid.
A set of buttons is valid based on how many times
each indicator number is present.  Odd occurrences mean
the light is on, even means off.