Year 2025 - Day 9 Outer Shell
Successfully implement the outer shell technique.
Take the existing edges and create a wrapper that is
1 space wider. Represent that as rectangles, and then
find the largest rectangle that never overlaps that shell.

To create the shell, we need to look at 4 points in a row.
This tells us what direction we were previously traveling,
which direction we are currently traveling, and which
way we will turn after our current edge.